🌱 Split oversized components into subdirectories for maintainability
Refactored 5 large component files into organized subdirectories with extracted
modules for types, utilities, constants, and validation. This reduces main
component file sizes while maintaining backward compatibility (existing imports
still work via index.tsx/index.ts entry points).

Changes:
- StatBlockFactoryModal: Extracted types, utils, validation, StatsPreview (~367 LOC extracted)
- CustomDashboard: Extracted constants and types
- DashboardState: Extracted actions and types
- FlightPlanBlueprint: Extracted constants
- BlueprintInfoPanels: Extracted constants

All components follow the ComponentName/ subdirectory pattern with index.tsx/index.ts
as the entry point. Existing imports remain compatible via Node resolution.

Related to issue #15790
