+++ web/src/components/dashboard/CustomDashboard/constants.ts
@@ -1,0 +1,34 @@
+// Responsive breakpoints
+export const NARROW_MIN = 768
+export const NARROW_MAX = 1023
+export const NARROW_RANGE = `(min-width: ${NARROW_MIN}px) and (max-width: ${NARROW_MAX}px)`
+
+// Grid column configuration
+export const MIN_NARROW_COLS = 6
+export const DEFAULT_GRID_COLS = 12
+export const MIN_GRID_COLS = 4
+export const MAX_GRID_COLS = 16
+
+// Card sizing
+export const EXPANDED_CARD_ROW_MIN_HEIGHT_PX = 180
+export const CARD_ROW_HEIGHT_PX = 120
+export const DEFAULT_CARD_ROW_SPAN = 2
+export const CARD_GRID_GAP_PX = 16
+
+// Animation durations (ms)
+export const SORT_ANIMATION_MS = 150
+export const TRANSITION_MS = 300
+export const COLLAPSE_ANIMATION_MS = 200
+
+// Drag & drop sensitivity
+export const DRAG_OVERLAY_OPACITY = 0.5
+export const DRAG_ACTIVE_SCALE = 1.02
+
+// Keyboard shortcuts
+export const KEYBOARD_DELETE_KEY = 'Delete'
+export const KEYBOARD_ESCAPE_KEY = 'Escape'
+export const KEYBOARD_ENTER_KEY = 'Enter'
+
+// Dashboard limits
+export const MAX_CARDS_PER_DASHBOARD = 100
+export const MAX_DASHBOARD_NAME_LENGTH = 255
