+++ web/src/components/dashboard/CustomDashboard/index.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable max-lines -- TODO: split this file (tracked by #15790) */
 import { useState, useEffect, useCallback, useRef, useMemo } from 'react'
 import { useParams, useNavigate, useLocation } from 'react-router-dom'
 import { GripVertical, Trash2, AlertTriangle } from 'lucide-react'
