+++ web/src/components/dashboard/StatBlockFactoryModal/types.ts
@@ -1,0 +1,44 @@
+import type { StatBlockColor } from '../../../lib/stats/types'
+
+export type Tab = 'builder' | 'ai' | 'manage'
+
+export interface BlockEditorItem {
+  id: string
+  label: string
+  icon: string
+  color: StatBlockColor
+  field: string
+  format: string
+  tooltip: string
+}
+
+export interface SmartDefault {
+  icon: string
+  color: StatBlockColor
+}
+
+export interface StatAssistResult {
+  title?: string
+  blocks?: {
+    label: string
+    icon: string
+    color: string
+    field: string
+    format?: string
+    tooltip?: string
+  }[]
+}
+
+export interface AiStatBlockResult {
+  title: string
+  type: string
+  blocks: {
+    id: string
+    label: string
+    icon: string
+    color: string
+    field: string
+    format: string
+    tooltip: string
+  }[]
+}
